+++ src/MySql/test/Update-AzMySqlFlexibleServerConfiguration.Tests.ps1
@@ -13,13 +13,13 @@ while(-not $mockingPath) {
 
 Describe 'Update-AzMySqlFlexibleServerConfiguration' {
         It 'UpdateExpanded' {
-        $config = Update-AzMySqlFlexibleServerConfiguration -Name net_retry_count -ResourceGroupName $env.resourceGroup -ServerName $env.serverName -Value 15 -Source user-override
+        $config = Update-AzMySqlFlexibleServerConfiguration -Name net_retry_count -ResourceGroupName $env.resourceGroup -ServerName $env.flexibleServerName -Value 15 -Source user-override
         $config.Value | Should -Be 15
         $config.DefaultValue | Should -Be 10
     }
 
     It 'UpdateViaIdentityExpanded' {
-        $ID = "/subscriptions/$($env.SubscriptionId)/resourceGroups/$($env.resourceGroup)/providers/Microsoft.DBForMySql/flexibleServers/$($env.serverName)/configurations/wait_timeout"
+        $ID = "/subscriptions/$($env.SubscriptionId)/resourceGroups/$($env.resourceGroup)/providers/Microsoft.DBForMySql/flexibleServers/$($env.flexibleServerName)/configurations/wait_timeout"
         $config = Update-AzMySqlFlexibleServerConfiguration -InputObject $ID -Value 150 -Source user-override
         $config.Value | Should -Be 150
         $config.DefaultValue | Should -Be 28800