+++ tools/Az.Tools.Predictor/Az.Tools.Predictor.Test/Mocks/MockAzPredictorTelemetryClient.cs
@@ -12,8 +12,8 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using System;
 using Microsoft.Azure.PowerShell.Tools.AzPredictor.Telemetry;
+using System;
 
 namespace Microsoft.Azure.PowerShell.Tools.AzPredictor.Test.Mocks
 {
@@ -55,6 +55,11 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor.Test.Mocks
 
         /// <inheritdoc/>
         public void OnGetSuggestion(GetSuggestionTelemetryData telemetryData)
+        {
+        }
+
+        /// <inheritdoc/>
+        public void OnLoadParameterMap(ParameterMapTelemetryData telemetryData)
         {
         }