+++ tools/Az.Tools.Predictor/Az.Tools.Predictor/ParameterSet.cs
@@ -24,6 +24,9 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor
     /// does not matter to resulting prediction - the prediction should adapt to the
     /// order of the parameters typed by the user.
     /// </summary>
+    /// <remarks>
+    /// This doesn't handle the positional parameters yet.
+    /// </remarks>
     sealed class ParameterSet
     {
         /// <summary>
@@ -36,11 +39,14 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor
             Validation.CheckArgument(commandAst, $"{nameof(commandAst)} cannot be null.");
 
             var parameters = new List<Parameter>();
-            var elements = commandAst.CommandElements.Skip(1);
             CommandParameterAst param = null;
             Ast arg = null;
-            foreach (Ast elem in elements)
+
+            // Loop through all the parameters. The first element of CommandElements is the command name, so skip it.
+            for (var i = 1; i < commandAst.CommandElements.Count(); ++i)
             {
+                var elem = commandAst.CommandElements[i];
+
                 if (elem is CommandParameterAst p)
                 {
                     AddParameter(param, arg);
@@ -68,11 +74,22 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor
 
             Parameters = parameters;
 
-            void AddParameter(CommandParameterAst parameterName, Ast parameterValue)
+            void AddParameter(CommandParameterAst parameter, Ast parameterValue)
             {
-                if (parameterName != null)
+                if (parameter != null)
                 {
-                    parameters.Add(new Parameter(parameterName.ParameterName, (parameterValue == null) ? null : CommandLineUtilities.UnescapePredictionText(parameterValue.ToString())));
+                    var value = parameterValue?.ToString();
+                    if (value == null)
+                    {
+                        value = parameter.Argument?.ToString();
+                    }
+
+                    if (value != null)
+                    {
+                        value = CommandLineUtilities.UnescapePredictionText(value);
+                    }
+
+                    parameters.Add(new Parameter(parameter.ParameterName, value));
                 }
             }
         }