+++ src/Synapse/Synapse/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 0.7.0
+* Simplify `Restore-AzSynapseSqlPool` cmdlet to make it consistent with the existing SQL DW cmdlet
+
 ## Version 0.6.0
 * Added support for operation of Advanced Threat Protection settings in SqlPool-level
     - Add `Update-AzSynapseSqlPoolAdvancedThreatProtectionSetting` cmdlet